requirements.txt
flask
pandas
openpyxl
google-genai
gunicorn